Crate zwave_tx

Modules

zwave_tx

Structs

_S_ROUTE_LINK_

rssi_val

RSSI value array used in zwapi_tx_report_t. Each value is an RSSI feedback constant defined above.

zwapi_tx_report

@brief Detailed report and data about Z-Wave transmissions

zwave_controller_connection_info

@brief Structure holding information about the source and destination when transmitting and receiving Z-Wave frames.

zwave_tx_options

Enums

E_ROUTING_SCHEME

Transport routing scheme state define definitions

_E_ROUTING_SCHEME_

Transport routing scheme state define definitions

rf_power_level_t

RF power level values used with zwapi_set_rf_power_level() and zwapi_get_rf_power_level()

Constants

ZWAVE_TX_QOS_MAX_PRIORITY

ZWAVE_TX_QOS_MIN_PRIORITY

ZWAVE_TX_QOS_RECOMMENDED_GET_ANSWER_PRIORITY

ZWAVE_TX_QOS_RECOMMENDED_NODE_INTERVIEW_PRIORITY

ZWAVE_TX_QOS_RECOMMENDED_POLLING_PRIORITY

ZWAVE_TX_QOS_RECOMMENDED_TIMING_CRITICAL_PRIORITY

ZWAVE_TX_RECOMMENDED_QOS_GAP

Functions

zwave_tx_abort_transmission⚠

@brief Abort a queued or ongoing transmission.

zwave_tx_fixt_setup⚠

Setup fixture for Z-Wave TX.

zwave_tx_init⚠

@brief Initialize the zwave_tx component.

zwave_tx_log_queue⚠

@brief Log the contents of the TX Queue.

zwave_tx_send_data⚠

@brief Queue and send frames to Z-Wave nodes.

zwave_tx_send_test_frame⚠

@brief Queues and send a test frame to a Z-Wave node.

zwave_tx_shutdown⚠

@brief Stop the tx component.

Type Definitions

S_ROUTE_LINK

on_zwave_tx_send_data_complete_t

@brief Callback function of zwave_tx_send_data(), indicating the result of the operation

sl_status_t

zwapi_tx_report_t

@brief Detailed report and data about Z-Wave transmissions

zwave_controller_connection_info_t

@brief Structure holding information about the source and destination when transmitting and receiving Z-Wave frames.

zwave_endpoint_id_t

zwave_multicast_group_id_t

zwave_node_id_t

Z-Wave NodeID type

zwave_tx_options_t

zwave_tx_session_id_t

@brief Handle that can be used for aborting ongoing transmissions or identifying TX Queue elements.