Function zwave_tx::zwave_tx_abort_transmission

pub unsafe extern "C" fn zwave_tx_abort_transmission(
    session: zwave_tx_session_id_t
) -> sl_status_t

@brief Abort a queued or ongoing transmission.

Calling this function will attempt to cancel/abort a queued transmission. If the transmission data has been passed to the Z-Wave module, it means that no further route resolution will be made, however the frame may still get transmitted and the callback of \ref zwave_tx_send_data will still be called. As a consequence the transmit status of on_zwave_tx_send_data_complete_t can be both TRANSMIT_COMPLETE_OK and TRANSMIT_COMPLETE_NO_ACK

@param session Session id returned by zwave_tx_send_data

@returns

SL_STATUS_OK if the frame transmission has been aborted (removed from the queue and not sent)

SL_STATUS_NOT_FOUND if the session_id is not valid.

SL_STATUS_IN_PROGRESS if the frame transmission was initiated and aborted. The on_send_complete callback will be called.