Function zwave_tx::zwave_tx_send_data

pub unsafe extern "C" fn zwave_tx_send_data(
    connection: *const zwave_controller_connection_info_t, 
    data_length: u16, 
    data: *const u8, 
    tx_options: *const zwave_tx_options_t, 
    on_send_complete: on_zwave_tx_send_data_complete_t, 
    user: *mut c_void, 
    session: *mut zwave_tx_session_id_t
) -> sl_status_t

@brief Queue and send frames to Z-Wave nodes.

This function is used to transmitting Z-Wave frames. The provided payload will be encapsulated into the format described in connection->encapsulation. This module does not verify if the destination actually supports the requested encapsulation, or if the destination exists at all (NodeID / Endpoint).

This module will put all frames on a queue. The frames with the highest qos_priority will be send first. the tx_options and connection parameters allow all kind of frames to be transmitted, including multicast/broadcast.

@param connection Connection object describing the source and destination. If either the source or destination endpoints ID are not null, the frame will be Multi Channel encapsulated. @param data_length Length of the frame to send @param data Points to the payload to send @param tx_options Transmit options to use. @param on_send_complete Callback function that will be called when the send operation has completed @param user User pointer passed in argument of the on_send_complete callback function @param session Pointer to location where to write the session id of the queued message. If this is set NULL the session id will not be written

@returns

SL_STATUS_OK The transmission request has been accepted and callback will be triggered when the operation is completed.

SL_STATUS_FAIL The transmission request has been rejected, callback will not be called.