Function zwave_tx::zwave_tx_send_test_frame

pub unsafe extern "C" fn zwave_tx_send_test_frame(
    destination_node_id: zwave_node_id_t, 
    power_level: rf_power_level_t, 
    on_send_complete: on_zwave_tx_send_data_complete_t, 
    user: *mut c_void, 
    session: *mut zwave_tx_session_id_t
) -> sl_status_t

@brief Queues and send a test frame to a Z-Wave node.

This function is used for sending a special test frame that will be testing the direct range connectivity of a remote Z-Wave Node.

Test frame are automatically send with the lowest priority.

@param destination_node_id Z-Wave NodeID of the destination. @param power_level Powerlevel setting to apply for the transmission. Refer to \ref rf_power_level_t @param on_send_complete Callback function that will be called when the send operation is completed @param user User pointer passed in argument of the on_send_complete callback function @param session Pointer to location where to write the session id of the queued message. If this is set NULL the session id will not be written

@returns

SL_STATUS_OK The transmission request has been accepted and callback will be triggered when the operation is completed.

SL_STATUS_FAIL The transmission request has been rejected, callback will not be called.